requirements.txt
fastapi
uvicorn
tensorflow-cpu==2.20.0rc0
numpy